+++ encrypted_number.cpp
@@ -1,7 +1,7 @@
 #include "encrypted_number.h"
 
-EncryptedNumber::EncryptedNumber(PaillierPublicKey* public_key, int ciphertext,
-                                 int exponent) {
+EncryptedNumber::EncryptedNumber(PaillierPublicKey* public_key, long ciphertext,
+                                 long exponent) {
     this->public_key = public_key;
     this->_ciphertext = ciphertext;
     this->exponent = exponent;
@@ -17,7 +17,7 @@ EncryptedNumber EncryptedNumber::operator+(EncodedNumber& other) {
     return this->_add_encoded(other);
 }
 
-EncryptedNumber EncryptedNumber::operator+(int other) {
+EncryptedNumber EncryptedNumber::operator+(long other) {
     return this->_add_scalar(other);
 }
 
@@ -27,12 +27,12 @@ EncryptedNumber EncryptedNumber::operator+(float other) {
 
 // Multiplying
 EncryptedNumber EncryptedNumber::operator*(EncodedNumber& other) {
-    int product = this->_raw_mult(other.encoding);
-    int exponent = this->exponent + other.exponent;
+    long product = this->_raw_mult(other.encoding);
+    long exponent = this->exponent + other.exponent;
     return EncryptedNumber(this->public_key, product, exponent);
 }
 
-EncryptedNumber EncryptedNumber::operator*(int other) {
+EncryptedNumber EncryptedNumber::operator*(long other) {
     EncodedNumber encoding = EncodedNumber::encode(this->public_key, other);
     return this->operator*(encoding);
 }
@@ -43,7 +43,7 @@ EncryptedNumber EncryptedNumber::operator*(float other) {
 }
 
 // Division
-EncryptedNumber EncryptedNumber::operator/(int scalar) {
+EncryptedNumber EncryptedNumber::operator/(long scalar) {
     return this->operator*(1.0f / scalar);
 }
 
@@ -51,7 +51,7 @@ EncryptedNumber EncryptedNumber::operator/(float scalar) {
     return this->operator*(1 / scalar);
 }
 
-int EncryptedNumber::ciphertext(bool be_secure) {
+long EncryptedNumber::ciphertext(bool be_secure) {
     if (be_secure && this->_is_obfuscated) {
         this->obfuscate();
     }
@@ -59,21 +59,21 @@ int EncryptedNumber::ciphertext(bool be_secure) {
     return this->_ciphertext;
 }
 
-EncryptedNumber EncryptedNumber::decrease_exponent_to(int new_exp) {
+EncryptedNumber EncryptedNumber::decrease_exponent_to(long new_exp) {
     if (new_exp > this->exponent) {
         throw std::runtime_error(std::string("New exponent should be smaller than old exponent."));
     }
 
     EncryptedNumber multiplied =
-        this->operator*((int)pow(BASE, this->exponent - new_exp));
+        this->operator*((long)pow(BASE, this->exponent - new_exp));
     multiplied.exponent = new_exp;
 
     return multiplied;
 }
 
 void EncryptedNumber::obfuscate() {
-    int r = this->public_key->get_random_lt_n();
-    int r_pow_n = powmod(r, this->public_key->n, this->public_key->nsquare);
+    long r = this->public_key->get_random_lt_n();
+    long r_pow_n = powmod(r, this->public_key->n, this->public_key->nsquare);
     this->_ciphertext = this->_ciphertext * r_pow_n % this->public_key->nsquare;
     this->_is_obfuscated = true;
 }
@@ -91,8 +91,8 @@ EncryptedNumber EncryptedNumber::_add_encoded(EncodedNumber& encoded) {
         b = b.decrease_exponent_to(a.exponent);
     }
 
-    int encrypted_scalar = a.public_key->raw_encrypt(b.encoding, 1);
-    int sum_ciphertext = a._raw_add(a.ciphertext(false), encrypted_scalar);
+    long encrypted_scalar = a.public_key->raw_encrypt(b.encoding, 1);
+    long sum_ciphertext = a._raw_add(a.ciphertext(false), encrypted_scalar);
 
     return EncryptedNumber(a.public_key, sum_ciphertext, a.exponent);
 }
@@ -110,12 +110,12 @@ EncryptedNumber EncryptedNumber::_add_encrypted(EncryptedNumber& other) {
         b = b.decrease_exponent_to(a.exponent);
     }
 
-    int sum_ciphertext = a._raw_add(a.ciphertext(false), b.ciphertext(false));
+    long sum_ciphertext = a._raw_add(a.ciphertext(false), b.ciphertext(false));
 
     return EncryptedNumber(a.public_key, sum_ciphertext, a.exponent);
 }
 
-EncryptedNumber EncryptedNumber::_add_scalar(int scalar) {
+EncryptedNumber EncryptedNumber::_add_scalar(long scalar) {
     EncodedNumber encoded = EncodedNumber::encode(this->public_key, scalar, this->exponent);
     return this->_add_encoded(encoded);
 }
@@ -124,18 +124,18 @@ EncryptedNumber EncryptedNumber::_add_scalar(float scalar) {
     return this->_add_encoded(encoded);
 }
 
-int EncryptedNumber::_raw_add(int e_a, int e_b) {
+long EncryptedNumber::_raw_add(long e_a, long e_b) {
     return e_a * e_b % this->public_key->nsquare;
 }
 
-int EncryptedNumber::_raw_mult(int plaintext) {
+long EncryptedNumber::_raw_mult(long plaintext) {
     if (plaintext < 0 || plaintext >= this->public_key->max_int) {
         throw std::runtime_error(std::string("Scalar out of bounds."));
     }
 
     if (this->public_key->n - this->public_key->max_int <= plaintext) {
-        int neg_c = invert(this->ciphertext(false), this->public_key->nsquare);
-        int neg_scalar = this->public_key->n - plaintext;
+        long neg_c = invert(this->ciphertext(false), this->public_key->nsquare);
+        long neg_scalar = this->public_key->n - plaintext;
         return powmod(neg_c, neg_scalar, this->public_key->nsquare);
     }
