+++ public_key.h
@@ -11,13 +11,13 @@ class PaillierPublicKey {
     long g;
     PaillierPublicKey(long n);
     bool operator==(const PaillierPublicKey& other);
-    int get_random_lt_n();
-    int raw_encrypt(int plaintext, int r_value = 0);
+    long get_random_lt_n();
+    long raw_encrypt(long plaintext, long r_value = 0);
     
-    EncryptedNumber encrypt(EncodedNumber value, float precision = 0, int r_value = 0);
-    EncryptedNumber encrypt(int value, float precision = 0, int r_value = 0);
-    EncryptedNumber encrypt(float value, float precision = 0, int r_value = 0);
+    EncryptedNumber encrypt(EncodedNumber value, float precision = 0, long r_value = 0);
+    EncryptedNumber encrypt(long value, float precision = 0, long r_value = 0);
+    EncryptedNumber encrypt(float value, float precision = 0, long r_value = 0);
 
-    EncryptedNumber encrypt_encoded(EncodedNumber encoding, int r_value);
+    EncryptedNumber encrypt_encoded(EncodedNumber encoding, long r_value);
 };
 #endif
